requirements.txt
python-docx==1.1.2
python-bidi==0.6.3
requests==2.32.3